Removing "namespace.twig" template, renaming other templates.
The "namespace.twig" template is being removed as it is not really used. Plus it has the potential to cause file name conflicts with classes that share the same name as the namespace. I instead opted to use the other index templates, which have been numbered so that they can be ordered and prevent that naming conflict.